refactor(event-handler): collapse createLambdaHandler into a Transport hook
createHandler and createLambdaHandler were near-identical copies of the same
loop; Lambda forked the whole thing only to bind the raw event + context into
the request container. Introduce a Transport abstraction in event-handler-core
whose sole job is the transport-specific "extract" step (bind raw platform args
into the container). createHandler now takes an optional transport (default
no-op), is variadic, and runs transport.extract(child, ...rawArgs) before
request setup; matching still uses rawArgs[0]. The richer "no event type
matched" diagnostic moves into core so every transport benefits.

createLambdaHandler becomes a thin delegation over createHandler with an
awsLambdaTransport that binds AwsLambdaEvent/AwsLambdaContext. Node's
createNodeHandler already delegated and needs no change (no extract step).
No consumer changes: CreateLambdaHandlerOptions and the (event, context)
signature are preserved.
